게시글 생성, 삭제, 수정 API 유저 검증 추가

+++ server/controllers/board.js
@@ -3,9 +3,25 @@ const models = require('../models');
 module.exports = {
   createPost: async (req, res) => {
     try {
-      const { body } = req;
-      const { id, title, content } = body;
-      const data = await models.board.createPost({ id, title, content });
+      const { body, decoded } = req;
+      const { title, content } = body;
+      const { userId } = decoded;
+      if (title.trim() === '') {
+        res.status(400).send('제목을 입력해주세요');
+        return;
+      }
+
+      const [user] = await models.user.findByUserId({ userId });
+      if (!user) {
+        res.status(404).send('존재하지 않는 유저');
+        return;
+      }
+
+      const data = await models.board.createPost({
+        id: user.id,
+        title,
+        content,
+      });
       res.status(200).send({ id: data.insertId });
     } catch (error) {
       console.log(error);
@@ -30,11 +46,24 @@ module.exports = {
   },
   updatePost: async (req, res) => {
     try {
-      const { body, params } = req;
-      const { author, title, content } = body;
+      const { body, params, decoded } = req;
+      const { title, content } = body;
       const { id } = params;
+      const { userId } = decoded;
+
+      if (title.trim() === '') {
+        res.status(400).send('제목을 입력해주세요');
+        return;
+      }
+
+      const [user] = await models.user.findByUserId({ userId });
+      if (!user) {
+        res.status(404).send('존재하지 않는 유저');
+        return;
+      }
+
       const [post] = await models.board.findPostById({ id });
-      if (post.author !== author) {
+      if (post.userId !== user.id) {
         res.status(403).send('유효하지 않은 요청');
         return;
       }
@@ -47,8 +76,22 @@ module.exports = {
   },
   deletePost: async (req, res) => {
     try {
-      const { params } = req;
+      const { params, decoded } = req;
       const { id } = params;
+      const { userId } = decoded;
+
+      const [user] = await models.user.findByUserId({ userId });
+      if (!user) {
+        res.status(404).send('존재하지 않는 유저');
+        return;
+      }
+
+      const [post] = await models.board.findPostById({ id });
+      if (post.userId !== user.id) {
+        res.status(403).send('유효하지 않은 요청');
+        return;
+      }
+
       await models.board.deletePost({ id });
       res.status(200).send('게시글 삭제 성공');
     } catch (error) {
